Include all testcases statically
Summary: Before, we captured CARGO_MANIFEST_DIR at compile time, but then used the value of it at runtime. That meant you couldn't move the binary around, which sucked. Instead, capture all the parsing test cases at compile time. Sadly, that means having a list of them in the file, but it's not too onerous, and never changes, so not a big deal.

Reviewed By: blackm00n

Differential Revision: D26036366

fbshipit-source-id: eaf3adea0a653b838a5c36e3a08ec7486a7a356c