Make AstStmt not imply Box
Summary:
Looking around, there are plenty of places we could optimise Box usage, but for AstStmt, the Box is implicit, and we'd rather not write Spanned<Stmt>. So make the Box explicit. Doesn't yet remove or reducing boxing in any places.

The only interesting diff is:

```
- pub type AstStmt = Box<Spanned<Stmt>>;
+ pub type AstStmt = Spanned<Stmt>;
```

Everything else is an inevitable consequence of that.

Note that we ignore two additional Clippy warnings where Clippy suggests we should be using less Box. That's because Clippy can see more, and we can now go back and clean it up after (which I do further up the stack).

Reviewed By: cjhopman

Differential Revision: D26100766

fbshipit-source-id: a97968ce1e719220870ce70507575d5b7da78e64