Make AstExpr not imply Box
Summary:
Looking around, there are plenty of places we could optimise Box usage, but for AstExpr, the Box is implicit, and we'd rather not write Spanned<Expr>. So make the Box explicit. Doesn't yet remove or reducing boxing in any places.

The only interesting diff is:

```
- pub type AstExpr = Box<Spanned<Expr>>;
+ pub type AstExpr = Spanned<Expr>;
```

Everything else is an inevitable consequence of that.

Reviewed By: cjhopman

Differential Revision: D26100767

fbshipit-source-id: 15d7aa8772c4f79c6b434448f8351c96daeb40a3