Make js.star valid Starlark
Summary: This test case was invalid Starlark. I thought about modifying our definition of Starlark to match what it did - but that's not a good path to go down. There is a spec, this code is clearly violating it, so fix the code, note it, and switch to testing everything.

Reviewed By: bobyangyf

Differential Revision: D26123954

fbshipit-source-id: ebdcb51bc132e097ed4e7805a71c267576f7e546