Add a lint for duplicate dictionary keys
Summary: `{1: 1, 1: 2}` is bad practice. It's disallowed by Go Starlark, but allowed by Python and us. I'm on the fence about whether it should be an error - it's better to be closer to Python, and you could imagine patterns where it makes code simpler. But most of the time its not what people will be expecting. For the moment, add a linter, which is enough to make me think we can punt on a decision for now.

Reviewed By: bobyangyf

Differential Revision: D26123955

fbshipit-source-id: 1e15bffd8f55c5fcd91b9f68fe2e26aac139e92d