Ban 07 to mean octal digits
Summary:
In both Python and Starlark numbers with leading zeros are banned because of the confusion with implicit octal notation. Our implementation actually used implicit octals, which was not spec compliant, and is pretty dangerous. Fix that.

I would have liked to write a lexer test directly, but we don't support those, so I added a grammar test instead.

Reviewed By: cjhopman

Differential Revision: D26185024

fbshipit-source-id: 7eb3c582f7d46b4df58318cba80fbdd7ba50de38