Remove the "not in" keyword
Summary: The NotIn keyword coming out of the lexer is odd. It causes problems with some lexer generators, it's not a thing in Python3 grammar. It has a variable amount of embedded whitespace. Far easier to delete all the special cases associated with it, and move NotIn as two separate keywords to the grammar. Simpler and seems to work just the same.

Reviewed By: bobyangyf

Differential Revision: D26183908

fbshipit-source-id: 5dcd72963d1ee4b18a3a765d952934da3704f1c5