Make Call evaluate its arguments in the right order
Summary:
Previously, if the user wrote `fun(a=f(), *[g()])` we would evaluate `g` before `f`, which is observable if the arguments have side effects. That violates the Starlark spec, and is confusing. The reason was that in Call we stored the positional, named, args and kwargs int that order. Now, we instead store them as a Vec in whichever order they were defined.

Making that change means that ParameterParser can now be called with positional (as a result of args) after named. That showed up a bug where we need to check that the thing we are replacing with positional hasn't already been set, using an is_assigned check. Benchmarking on a large call-heavy benchmark that unassigned check reduces the performance from 50s to 60s. To regain the performance I added a flag which tracks at runtime whether we've seen a named flag yet or not, which regains the performance back to 50s.

Reviewed By: cjhopman

Differential Revision: D26172781

fbshipit-source-id: 74ebc1f0343c92449cb27feed89ed15f49631b32