Split out the JSON lint type from the normal Message type
Summary: We want to generate lint messages with --json in a predictable and consistent type. We also want a good Message type to represent all messages, which doesn't want a code, but does want a span rather than a position. In order to avoid breaking JSON messages in future, have a separate type for them.

Reviewed By: blackm00n

Differential Revision: D26313848

fbshipit-source-id: 8ebdc5810a876114c8683926839c83b3f0a75156