Add a new lexer
Summary:
The older lexer was pretty slow, and manually coded. I've written a new lexer based on the `logos` library. Initial benchmarking on a representative file shows:

* Old: 1.275ms
* New: 0.479ms
* New without allocation: 0.188ms

The old/new are obvious, but new without allocation indicates what we can get to if instead of using String we use &str - which in most cases is easy with the new lexer.

Since this is a big change, and the behaviour of the lexer is a little hard to fathom (part of why I want to redo it), this patch lands turning on both the old and the new lexer, with equality checks on their output. As a consequence, if the lexers differ, it will throw a runtime error. In our tests that doesn't happen. I'm sending it through CI like that, and think that is worth a discrete commit, then in the next diff I'll delete the old lexer.

There are various quirks of the old lexer that I have replicated precisely, like whether it outputs a newline at the end of a file or not. Those are irrelevant details, and I'll strip out most of that logic once the old lexer is done.

In going through the old lexer, I found bugs on paths that weren't tested. I fixed one or two, but I'll go through and do additional testing after we're down to only one lexer.

Reviewed By: milend

Differential Revision: D26369385

fbshipit-source-id: 4090975859a6c4391d175dcd9f7ff05433cd56f7