Create fresh invalid name errors
Summary:
Previously we created an error about an invalid name, then cloned it. However, cloning errors is a pretty rare thing to do, because anyhow::Error doesn't support it. In this case it's easy to just create a fresh error each time, so do that.

Interesting, I tried doing this a number of months ago, and got caught up in Rust lifetime errors and failed entirely. Fortunately, this time it wasn't too difficult (and I don't believe it was any harder in code than last time).

Reviewed By: bobyangyf

Differential Revision: D26453489

fbshipit-source-id: 9b5e54b7e0586fd0077e9bd454e8147815134460