Avoid duplicating source file strings
Summary:
Now we take the Starlark string to parse as a String, rather than a str which we clone. Avoids copying the file in many cases.

For testing, I normally just do a clone in the test harness, since that makes writing the test case easier, and we don't care about redundant copies there.

Reviewed By: bobyangyf

Differential Revision: D26453464

fbshipit-source-id: 355358dc977d2e91496796bda72dfb584c85cca9