Get rid of the PartialEq constraint on LexerError
Summary: Can't be available on anyhow::Error. Weaken the two tests that use it (unterminated string literals), but those tests weren't that good (a lot of ways the code could be wrong without failing the test), and we'll need to rewrite them if we move all the lexer tests to anyhow. Accepting a temporary regression in the test suite while I rewrite this bit.

Reviewed By: bobyangyf

Differential Revision: D26453480

fbshipit-source-id: a9e1abc9d2984b46404340a7e1e5b366eeb15eee