Use assert_eq!(x.eq(&x), true) to check equality
Summary: When checking equality we don't want to write `assert_eq!(x, x)` as that's a very easy typo to write, and a Clippy lint. We also don't want to write `assert!(x == x)` as that should be written `assert_eq!` in most cases. The format `assert_eq!(x.eq(&x), true)` is pretty unusual otherwise, so makes it clear what we are doing is checking the equality instance.

Reviewed By: bobyangyf

Differential Revision: D26483804

fbshipit-source-id: e006029e116dd656b38ae9387c78d7bb7fdcdc04