Don't have compare default to Equal if pointers are equal
Summary: Before, for types that weren't comparable, e.g. dict, {} < {} would return True if the pointers were equal, but error otherwise. That doesn't follow the Starlark standard, and means that whether a program fails or not depends on subtle details like whether definitions manage to get shared. Instead, push information about whether pointers are equal down to the implementor of compare, with the default throwing an error.

Reviewed By: cjhopman

Differential Revision: D26576978

fbshipit-source-id: 228cf9690fcfa0fe13a6337a93ea36519d6b91f6