Improve the break/return/continue/load validation
Summary: Previously we didn't check that return couldn't occur inside a function because it was a parse error. We also didn't check that load was at the top-level. In preparation for relaxing that restriction, make the static validity checker actually check all the conditions that can arise given the whole AST.

Reviewed By: cjhopman

Differential Revision: D26576334

fbshipit-source-id: d5c09e2acc8b64e9c13ca1073d62ee1eb0c87dc1