Don't use pointer equality short-circuit in comparison
Summary: Previously, if we did `a < b` then we first checked if they were identical pointers, to short-circuit things. That means `None < None`, which is invalid, so we moved to first doing a type test before short-circuiting. But that's still invalid as `(None,) < (None,)` if the pointers are equal, but not otherwise. Let's just give up on pointer equality for comparisons, because comparisons aren't particularly common, and pointer equality is only safe if the type definitely only contains comparable types, which is just too difficult to compute safely.

Reviewed By: bobyangyf

Differential Revision: D26732954

fbshipit-source-id: b828377d7a0ec8e8432b4ed940e31f5398bcb78a