Hide dependencies of exported macro_rules from users
Summary:
Currently, we use a `paste::` reference to refer to the paste crate.
In dependent crates, this appears just like that and then requires that that
dependent crate has its own paste dependency.

Instead, we can add a symbol in the starlark-rust crate itself pointing to
`paste` and then reference it through `$crate`.

Reviewed By: ndmitchell, dtolnay

Differential Revision: D26749910

fbshipit-source-id: 5e3e38126c8ecd6df2afe8531b68de2692084d22