Add a Deref impl for starlark RefFoo types
Summary:
starlark_value!() and starlark_immutable_value() generate RefFoo types
to make it easy to provide starlark functions that accept specific types. All
uses of that currently need to explicit reach into the `.0` member of the RefFoo
type, which seems unnecessary. This adds a Deref impl that derefs to the
wrapped reference.

Reviewed By: ndmitchell

Differential Revision: D26749920

fbshipit-source-id: b246c70be98d568b325d7a2c60b6912bf6dabd4e