Rename EvaluationContext to Evaluator
Summary:
What if we lived in a world where not everything in Rust had to be named Context? Start that process by renaming EvaluationContext to Evaluator, which is an equally good name, but shorter and less overloaded.

There's plenty of places we do `ctx = Evaluator::new()` now, which we can clean up over time - I have no immediate rush to clean them up now.

Reviewed By: bobyangyf

Differential Revision: D26912229

fbshipit-source-id: 94700ea56043cf5e0990a84c4c8675bddd1d7f14