Move Value doc tests to the trait
Summary: Previously for things like `+` we had doc tests for that in list, string and tuple. Doc tests are expensive in terms of execution time, and they were on traits, so its much harder to see the associated docs. Instead move the docs onto the trait, which makes them more visible, and fewer of them, which should slightly reduce test time.

Reviewed By: milend

Differential Revision: D27043375

fbshipit-source-id: d1b15c1992c095e1caa57a19b4af315e67351157