Add a test of Starlark GC
Summary:
Rather than trying to write lots of GC tests (which is not massively realistic), run all tests in three modes - GC never, auto and always. Gives the GC a very good kicking discovering a few bugs in the process.

Running the tests three times might have a negligible slowdown, but in truth its doc tests which run the Rust compiler which are by far the outliers, so I think this is not detectable.

Reviewed By: milend

Differential Revision: D27043219

fbshipit-source-id: 0ddfeb7511c1c85225a0e6293ed312af6128df2d