Simplify += for lists
Summary: The Starlark spec says += does one thing on lists, and a different thing otherwise. That's a bit confusing and non-logical, but has performance implications if it doesn't work that way, so OK. Instead of delegating this to the trait, and ending up with pretty complex logic with weird borrowing rules etc, let's put it in the interpreter for real.

Reviewed By: blackm00n

Differential Revision: D26946862

fbshipit-source-id: 9246a95c6227b70ec062f36ad3fddb6b1b6b85dc