hash() as per spec
Summary:
Spec is https://github.com/bazelbuild/starlark/blob/master/spec.md#hash

Just changing `get_hash()` to return `i32` won't be a simple change.  Addressing Dict via `Value` with string and via raw `&str` should behave the same:
```
    pub fn get(&self, key: Value<'v>) -> anyhow::Result<Option<Value<'v>>> {
        Ok(self
            .content
            .get_hashed(key.get_hashed()?.borrow())
            .copied()
            .map(ValueLike::to_value))
    }

    pub fn get_str(&self, key: &str) -> Option<Value<'v>> {
        self.content
            .get(&ValueStr(key))
            .copied()
            .map(ValueLike::to_value)
    }
```

So when addressing with `Value` we go via:
```
    fn get_hashed(self) -> anyhow::Result<Hashed<Self>> {
        Ok(Hashed::new_unchecked(
            SmallHashResult::new_unchecked(self.get_hash()?),
            self,
        ))
    }
```
and hash is
```
SmallHashResult::new_unchecked(self.get_hash()?),
```

When addressing with raw `&str`, wrapped in `ValueStr`:
```
    pub fn get<Q>(&self, key: &Q) -> Option<&V>
    where
        Q: Hash + Equivalent<K> + ?Sized,
        K: Eq,
    {
        self.get_hashed(BorrowHashed::new(key))
    }
```
leading to:
```
impl<'a, Q: ?Sized> BorrowHashed<'a, Q> {
    pub fn new(key: &'a Q) -> Self
    where
        Q: Hash,
    {
        Self::new_unchecked(SmallHashResult::new(key), key)
    }
```
leading to:
```
impl SmallHashResult {
    pub fn new<K: Hash + ?Sized>(key: &K) -> Self {
        let mut hasher = DefaultHasher::new();
        key.hash(&mut hasher);
        Self::new_unchecked(hasher.finish())
    }
```

So for correctness:
```
self.get_hash()
```

should equal

```
let mut hasher = DefaultHasher::new();
key.hash(&mut hasher);
hasher.finish()
```

that is like that right now because:

`Box<str>`:
```
    fn get_hash(&self) -> anyhow::Result<u64> {
        let mut s = DefaultHasher::new();
        hash_string_value(self.as_ref(), &mut s);
        Ok(s.finish())
    }
```

Dict related code:
```
impl<'a> Hash for ValueStr<'a> {
    fn hash<H: Hasher>(&self, state: &mut H) {
        hash_string_value(self.0, state)
    }
}
```

but if we change `get_hash` to algorithm as per spec, we won't be able to match it in `ValueStr` `Hash` implementation, the only possible solution is to introduce double hashing in `ValueLike::get_hashed` and treat `get_hash` as a key which is ugly. So approach when starlark `hash()` method might return hash with different semantics seems to be better.

Reviewed By: ndmitchell

Differential Revision: D26960268

fbshipit-source-id: a8fcdcd879b180e5caa3489b64bbdc4aaed55c2b