Improve the stack of the error message
Summary: Previously, if we called a native function, while it would go on the call stack, the caller would "record" the call stack after it got popped. That wasn't fatal, as often the message gave a good clue as to what the call was, and usually highlighted it perfectly. But it's still not as it should be, so record the call stack earlier.

Reviewed By: bobyangyf

Differential Revision: D27229531

fbshipit-source-id: 047b84762fe7f73314c26d6ec00b67e586b685a5