If extra_v is used, disable GC
Summary: If the user uses extra_v, they get to stash values in places we can't see them. If so, doing GC is dangerous, and disable it. We could have a special walk method to recover GC, but in most cases it isn't necessary, so leave that until we have a real problem.

Reviewed By: blackm00n

Differential Revision: D27244095

fbshipit-source-id: 4cb7d57b73ddb59595c58939890fa80e7f8cdc10