Add print to the stdlib
Summary: Most people should add print to their Starlark library when they are starting, and when they get more advanced, move it somewhere with better logging. However, we defaulted to not providing it, which made getting going harder. Since we can now exclude specific extensions, it's much easier to add print, and if people want to do something custom, they can just not add it to the Globals.

Reviewed By: cjhopman

Differential Revision: D27393675

fbshipit-source-id: ddea83dfc71cfdb0238e73b7a2f42805c3bc5953