Change how extended functions are added
Summary: Before, if you wanted "extra functions" in your Starlark that included dedupe, map and filter as a set. It made it hard to get exactly what you wanted, and to include/exclude specific things. Instead introduce a LibraryExtension enum that lists all the extensions, so you can pick and choose.

Reviewed By: cjhopman

Differential Revision: D27393717

fbshipit-source-id: 571b9cd8527fede6d197dd28fed6aec58ba1428e