Stop tests from relying on Debug of String
Summary: In the latest Rust nightly, the character `'` gets rendered as `'` in a String, whereas it used to be `\'`. The Debug implementation is considered unstable, so these kind of things can happen. Instead, switch to using our json to string code in unlex, which is mostly the same, but our code so well defined what it does.

Reviewed By: bobyangyf

Differential Revision: D27748476

fbshipit-source-id: a45507986372103a1d78af2d26651f000ac00d57