Provide a non-allocating set_attribute function
Summary: Coming up with this was a bit of a brain teaser, so keen to reuse that logic. You often want to use GlobalsStatic to set fields. If you are using them with get_members, you can define a function with #[starlark_attribute], but it allocates the result each time and is a bit verbose. This new function doesn't allocate and can be used compactly.

Differential Revision: D27829799

fbshipit-source-id: 98d4af917e43293d271618705d0768c3ec9f6942