Make string.elems/codepoints return opaque iterators
Summary: The spec says that they should be opaque, presumably so they can be optimised to not allocate the whole list. We now do that, although because char allocates on the heap (for now) the win probably isn't as big as it would be. Ensures we can optimise in future without breaking though, and that we follow the spec.

Reviewed By: milend

Differential Revision: D28057098

fbshipit-source-id: 0747d758a910afebab83eef96a2a00ff6897dbb5