Optimise record construction
Summary:
Previously we created all the parameters and the invoking function every time we created a record. Now we do it when we create the record type, and reuse it on each actual invocation. Profiling suggested this could be a bottleneck.

The downside is we don't have the name of the function being called at this point, so error messages are fractionally worse. But we do have the location of the call, which is probably the exact same information, so I don't feel too bad.

Reviewed By: blackm00n

Differential Revision: D28116867

fbshipit-source-id: 4ae716270b8126c3a3712cfecf588367efd31687