Only store the parameter names once
Summary:
Previously we stored the parameter names in the kinds Vec and the names Map. The only use of the value from nameswas for error messages, and you can reconstruct that from the values in Map, so do that. Saves reallocating every argument string twice.

The one behavioural change is that previously if a user defined the function with *my_args, we'd print error messages about the caller in terms of my_args, and now we go with *args. That is so rare, and the name after * is irrelevant, so this seems cleaner anyway. When using #[starlark_module] it was already impossible not to write args there.

Reviewed By: blackm00n

Differential Revision: D28116872

fbshipit-source-id: 71bc914ff635addec0d35abb01e138d8d9e83d70