Generalise profiling to impl Write
Summary: Previously we had to write a file when producing the profile. Generalise it to work with any Write trait, so we can write to a Vec if we want (e.g. for testing). Don't expose the richer API to users though, since the format of the output is not yet stable.

Reviewed By: milend

Differential Revision: D28223357

fbshipit-source-id: e84ba488e1053833ca464cc162d2b1855ff2e4b4