Add a test for heap allocate field defaults
Summary: We do GC stress tests in the Assert module, which catches a lot of bugs in walk implementations. However, in the tests for fields, we were only using Int (not on the heap) or literals (added to the FrozenHeap at compile time). Fix that with a heap allocated default, which before the fix in the previous diff, crashed.

Reviewed By: milend

Differential Revision: D28253690

fbshipit-source-id: 4812594fa41292cda925abbcad4224d686a535af