Don't mutate on_stmt
Summary:
The debugger sets on_stmt to break into each statement. It then disables breakpoints while doing evaluation of expressions requested in the hover etc, by clearing on_stmt. But that's a pretty dubious thing to do, since its state owned by the Evaluator, that I poked out for thise case. Instead of abusing on_stmt, add a dedicated bool to say we are evaluating and should stop.

I've actually used an int, as that's easier to match enter/exit without relying on the invariant that there is only one simultaneous evaluation. But there is only one because Evaluator isn't thread safe, but doing the int is simpler and more robust.

Reviewed By: milend

Differential Revision: D28286942

fbshipit-source-id: e7af48215086ff5a1234228b65d0441a3fe04ed6