Move parameter collection up one level
Summary: Before we had an enum, and each variant had an idential ParametersCollect inside it, so we often switched on that enum to find the collect. Now pull that collect out to the top level, so its always in the same place. Should reduce the overhead in pushing parameters slightly, and also cleaner code.

Reviewed By: milend

Differential Revision: D28304682

fbshipit-source-id: 50cce3f33bd7eb60c511baed3c5673662b8b2ca6