Hide unassigned variables from users
Summary: Unassigned variables should only be used to model one specific type of interaction - variables that have not yet been assigned. We'd like to do static analysis in future to figure out which variables are unassigned, which is vastly easier if external users of the API can't make things unassigned.

Reviewed By: milend

Differential Revision: D28346085

fbshipit-source-id: 4462c4b48700ace5fd147bfe6e735424ec0e757a