Change module value return
Summary:
We have the feature that:

```
x = 100 + 7
x
```

"returns" the `x` as the final value of the module. Previously that was done by making every statement return a value, and having the very final statements value be kept. That's complex and wasteful. Instead, just inject a fake return before we start compiling, and watch for that return. That means statements don't need to return values.

Reviewed By: swgillespie

Differential Revision: D28424083

fbshipit-source-id: 02af01ea1fe44f4202405cdf5ced980a8dc37ed2