Make the span always start at 0
Summary: Not sure why this was 1. It makes no sense. But since we always did an offset by the base index, I think it cancelled out in subtractions. Even so, it's vastly more simple to use 0, and now we can start to unwind the file.span property - which was redundant anyway.

Reviewed By: bobyangyf

Differential Revision: D28473978

fbshipit-source-id: 82fd196ccf615ecf8fd2948a6a6736b10f70d665