Use the name eval for the evaluator
Summary:
Previously, we had ctx being the Evaluator binding in a #[starlark_module]. The reason for that naming choice was that previously the evaluator was called context. As things stand today, we have the name ctx, context and evaluator for the Evaluator. I want to standardise on eval as the one true name.

Alas, this change breaks the API of starlark_module, but it seems worth it to get long-term consitency. I'd have loved to keep compatibility by defining both eval and ctx, but alas, mutable references can't be copied.

Reviewed By: cjhopman

Differential Revision: D28539168

fbshipit-source-id: 623a669f2ac76b32302af3ad816935daeba7b9fd