Remove Unassigned from Value
Summary: Previously Value had a special Unassigned value. But in most cases, we didn't expect Unassigned and just panic'd. Now that Option<Value> is just as cheap as Value, we can switch to using None for the unassigned value, which is handy since in most places we thought we might have unassigned, we had to do a check and convert it into None. Now the type system works for us, and tells us where those None values might lie.

Reviewed By: milend

Differential Revision: D28601006

fbshipit-source-id: d8c25863fd93940386702928ec2be442b4b628d0