Make the GC points explicit
Summary:
There are only a very small number of GC points, at the top-level of a module. Rather than check whether GC is required many times, insert explicit checks where it is plausible. As a result, we can remove the is_module_scope check.

There are three ways to implement the GC check.

1. At runtime, with a conditional, which is what we were previously doing.
2. At compile time, with two different code branches, which would double the amount of statement closures generated (probably not a big deal) and avoid any runtime costs.
2. At compile time, with a second closure for places where we might have a GC. Makes runtime of GC checks slightly higher, but that's not a big deal, as they must be linear in the length of the users code, and are likely to be significantly less than that. I went with this approach.

Reviewed By: milend

Differential Revision: D28635984

fbshipit-source-id: 2f2869867032b49f61049cc596a54286b02bdb5b