Allow set_module_variable_at_some_point even when not at the top-level
Summary: This function has a guard about module_scope because it used to be used to set top-level variables normally, but that stopped being the case long ago. Might as well let people do this whenever they want, as its not harmful when not at the top-level. Main reason for doing this is because we don't really want to track whether we are at the top-level or not.

Differential Revision: D28635988

fbshipit-source-id: 043e545bd610fbfda8fe70dc778e01676addaf62