Don't have assert directly invoke a GC
Summary: Previously assert had methods to force a GC here and now, but it wasn't super sure if that was safe or not, so it kind of guessed. Rather than do that, instead change the last_heap_size in a way that forces the next possible GC point to do a GC then. Should be mostly equivalent, but a lot less invasive.

Reviewed By: milend

Differential Revision: D28635986

fbshipit-source-id: 97ec0d932ec39771e0900f1b13189bc662b8dfb5