Move to a single slot vector
Summary:
Before each pile of local slots was a Vec on its own. Move to a model where we have a single slot vector, which avoids doing the reallocation, with a corresponding significant saving on the function call benchmark (2x faster).

~~In the process I broke the debugging functionality. This was always a bit of a hack, and no one is using it for real, so going to break it temporarily. I think once we have a better stack situation we'll be able to do it properly.~~

Update: I figured out a 1 line change that keeps debugging working, by not updating the slot base when moving into a native function.

Reviewed By: swgillespie

Differential Revision: D28640428

fbshipit-source-id: 3a5684ec1674703a038bb770209660b936c27790