Don't make NativeFunction be type generic
Summary:
Despite warnings on the type that we should never make it generic because that's expensive, it doesn't really make any difference - which is expected, because we were making it generic before actually calling it anyway. Less code generated and easier to profile.

Benchmarks suggest this is either flat, or slightly faster.

Reviewed By: swgillespie

Differential Revision: D28673228

fbshipit-source-id: f35acf71c3540ea9fc0204e4fb3d532e543b6e69