Move the kwargs out of line
Summary:
The FunctionInvoker structure was large, in a way that made the compiler do a lot of expensive stuff. The largest thing in that structure was a SmallMap, which is ~10 words. However, if we ever use the kwargs, it's to allocate them on the heap, where we require them to be in a Box anyway. So take them out of line in FunctionInvoker and put them in a Box (allocated only if they are used), which gets reused directly if they get allocated.

See the impact on benchmark.py, which is pretty nice.

Reviewed By: swgillespie

Differential Revision: D28705499

fbshipit-source-id: 77e36bdac8d35398ea563241cb43994e8168c500