Move to using a macro for the SmallMap iterators
Summary: Before we defined them by hand, which was verbose, but also meant that we couldn't define faster variants for things like collect. Given we have an enumeration, doing that switch once for the entire iterator is a much better idea, so a collection implementation is a very good plan. Copies the same pattern from VecMap, but modified to be a wrapper around 3 different iterators.

Reviewed By: krallin

Differential Revision: D28929424

fbshipit-source-id: aaca8746878531ba0db2575f1157b943f7e97010