Optimise getattr
Summary:
Previously getattr called StarlarkValue.get_attr. If it returned an Error, and there was a default, it mapped that into the default. That meant we were creating anyhow::Error values and discarding them. That's disasterously expensive, because the code in the error paths (deliberately) does a good-but-expensive job of producing error messages. In practice, I saw this because struct was pretty printing its value, for it to be thrown away.

Instead make get_attr return an Option, and synthesise the error message later if you need it (only two places).

Differential Revision: D28991919

fbshipit-source-id: 4dac62bba50c2ec3bb62bbd397b8880f4154c242