Optimise NativeAttribute
Summary:
Before NativeAttribute was a NativeFunction, which took a single argument, then went through all the parameter machinery to call it again with that single argument. Cut all that stuff out, making NativeAttribute a first class thing, which avoids a redundant trip through invoke.

In practice, probably doesn't make any performance difference because we have so few `#[attribute]` things, but it's simpler anyway.

Reviewed By: swgillespie

Differential Revision: D29403054

fbshipit-source-id: e8bd3c5cea6e73c62ea6d6a6c367a3019828ab02