Switch the 1pos benchmark to use abs
Summary: Len is a hot function, so optimising it is a very good idea. That means to find something that measures the overhead of a single native function call we need to pick something else. I initially went for `bool`, but it seems in Python that isn't native and is horribly slow. Of the remaining candidates, `abs` seems the best. This isn't 100% fair (Python has a virtual call since it has more types of number), but that single virtual call seems pretty small in the benchmark, with both Starlark and Python decreasing by similar margins.

Reviewed By: swgillespie

Differential Revision: D29426129

fbshipit-source-id: bb35dfc0e59132710f908e5f1a5e1d7ab3724a75