Make from_value first check if the value is frozen or not
Summary:
Instead of testing if it's frozen or not, we can guarantee that only frozen things go on the frozen heap cheaply. This should be a nice performance win in some cases.

A trick suggested by Chris, that wasn't applicable initially because we had copy-on-write values. Now we don't, it's safe.

Reviewed By: bobyangyf

Differential Revision: D29660124

fbshipit-source-id: 8d7688fade44e1c032098e16cbd09c37f6b3fe09