Fix string indexing
Summary: Previously if you did "foo"[100] we would attempt to skip over the first 100 bytes of the string foo, and see where we were. That's obviously a hideous idea, as the string isn't that long. Now we try to only skip the minimum of the length of "foo" and 100, ensuring we stay within bounds. Reported as GitHub issues #24 and #25.

Reviewed By: abrassel

Differential Revision: D29701167

fbshipit-source-id: eb3b3b35a406e7582da9fc41b0551071216d5bbc