Tidy up the fast_string fix
Summary: If the number of bytes is too large, we don't need to scan to see how large it is, we can immediately stop. We also should check the debug assertion on fast_string, so if we make the same mistake again, we error rather than walk over invalid memory.

Reviewed By: blackm00n

Differential Revision: D29711076

fbshipit-source-id: 062716e0b86f819ef83b3d48de804887563cc019