Make get_ref return an AValue
Summary: Previously get_ref returned a StarlarkValue. Now it returns an AValue, which is slightly richer than a StarlarkValue (but mostly in ways people can't see). The benefit is that we can avoid vtable adjustments by keeping the same vtable everywhere, and can start to move special-purpose methods on to AValue without requiring users update.

Reviewed By: bobyangyf

Differential Revision: D29820616

fbshipit-source-id: bbf6ed1d709578c3400aeb3145cf4c7e7de862e2