Get rid of ValueMem::Ref
Summary: A type that was only used when lambda variables are captured, which is almost never. Therefore instead of having a special representation on the heap, just use the normal AValue representation. Does involve one extra `Box`, but since this never happens, simplicity is preferred (we might even win by having less alternatives).

Reviewed By: bobyangyf

Differential Revision: D29832115

fbshipit-source-id: 9809234dece6d248f70ccb25110c95fdc4241a40