Switch to a variable-sized heap for FrozenModule
Summary:
Instead of storing Box pointers to the values in the Starlark heap, we now store the real values. Less memory allocation, less fragmentation - should all be a win.

Plan is to do this for the Value heap in a few diffs time, at which point I'll actually measure the impact.

Reviewed By: krallin

Differential Revision: D29875419

fbshipit-source-id: 7923e34e1c89af48fd50244502a57cd2885a0192