Switch to a variable size heap
Summary:
This converts the Value heap to be variable size, so the payload is stored inline in the heap. Basically the same recipe as for the frozen heap, but with the addition of the overwrite capability to Arena2.

The performance benefit isn't huge - compared to the start of the stack one interesting benchmark is 29.6 vs 27.6. Definitely noticable, but not as large as I hoped.

Reviewed By: krallin

Differential Revision: D29909376

fbshipit-source-id: 9771a6e6cc4b353ac2abd8a3f16f0b03e1cdceba