starlark_module: add `derive(StarlarkAttrs)`
Summary:
This diff implements a proc-macro that allows us to `#[derive(StarlarkAttrs)]`
on a struct, and expose all the struct fields to starlark.

I think this could be made better by automatically implementing the `_attr`
methods for `StarlarkValue` on types that derive this trait, but this is a fine
start.

Reviewed By: ndmitchell

Differential Revision: D29850551

fbshipit-source-id: c8c8995675c1e3bbaebad422362818df32f0dcaa