Make the Arena tests use StarlarkAny rather than String
Summary: The String type in Starlark has lots of weird quirks - it's currently a Box<str> because of performance reasons, but that's highly likely to change. Instead of relying on such details, use the comparatively simple StarlarkAny type and put a String in it for testing, which should behave most like a normal data type.

Reviewed By: krallin

Differential Revision: D29960660

fbshipit-source-id: 63549cd5b2197727cedbb1d5a6d89286aaadcd3b