Clean up FrozenRef Box<str> pieces
Summary: Before we had FrozenRef unpack str which produces a Box<str>. In most cases we would have preferred a direct str, so do that by default, and rename the method that produces a Box<str> and actually needs it. Less code now depends on the Box<str> property which isn't guaranteed.

Reviewed By: bobyangyf

Differential Revision: D29960659

fbshipit-source-id: 000a2c5018ddffa9aa583909ebb1549998ae159d