Add with_iterator
Summary:
The for_each signature was complex and didn't include size hints (great for avoiding allocations), while the iterator is expensive to allocate. Happily there is a with_iterator pattern that avoids all the costs, with a single extra virtual call in place of the iterator allocation (easily worth it). Introduce that.

Plan is introduce the new one, move everyone over, delete for_each.

Reviewed By: krallin

Differential Revision: D30011197

fbshipit-source-id: aab8c3bc6ddb841d7ced52fba6fe05327b5a3f4d