Make the heap profile use the Starlark-level types
Summary: When we had thaw-on-write, we would get both List and FrozenList on the heap, so it was useful to get the underlying Rust type_name to differentiate them. Now we only have non-frozen values the Starlark name is shorter and more friendly, while being unambiguous, and we can avoid post-processing it to clean it up.

Reviewed By: krallin

Differential Revision: D30040664

fbshipit-source-id: 851cbf80bd22b2d608a00c8b9bc076c8b7200015