Optimise flame profiling
Summary: When building a flame profile we don't disable GC (which keeps things faster), but that means we have to walk the entire historical call stack on each trace - which was taking the majority of the time. Solve that by caching each Value only once, and using an index. Provided the number of unique functions is low, this reduces the time significantly. In practice, reduced performing one flame graph from 70s to 30s (when 20s is the standard compilation time).

Reviewed By: krallin

Differential Revision: D30132958

fbshipit-source-id: d0150ec772691018a9477edbde2d98bf2971169e